TI matter-v1.2-ti-1.0-EA-1.0 release

+++ src/ble/tests/TestBtpEngine.cpp
@@ -20,14 +20,15 @@
 #include <cstdint>
 #include <numeric>
 
+#include <pw_unit_test/framework.h>
+
+#include <lib/core/StringBuilderAdapters.h>
 #include <lib/support/CHIPMem.h>
 #include <lib/support/logging/CHIPLogging.h>
 
 #define _CHIP_BLE_BLE_H
 #include <ble/BleLayer.h>
 #include <ble/BtpEngine.h>
-
-#include <gtest/gtest.h>
 
 using namespace chip;
 using namespace chip::Ble;
@@ -67,7 +68,7 @@ TEST_F(TestBtpEngine, HandleCharacteristicReceivedOnePacket)
     };
 
     auto packet0 = System::PacketBufferHandle::NewWithData(packetData0, sizeof(packetData0));
-    EXPECT_EQ(packet0->DataLength(), 5);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(5));
 
     SequenceNumber_t receivedAck;
     bool didReceiveAck;
@@ -81,7 +82,7 @@ TEST_F(TestBtpEngine, HandleCharacteristicReceivedTwoPacket)
     constexpr uint8_t packetData1[] = { to_underlying(BtpEngine::HeaderFlags::kEndMessage), 0x02, 0xff };
 
     auto packet0 = System::PacketBufferHandle::NewWithData(packetData0, sizeof(packetData0));
-    EXPECT_EQ(packet0->DataLength(), 5);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(5));
 
     SequenceNumber_t receivedAck;
     bool didReceiveAck;
@@ -89,7 +90,7 @@ TEST_F(TestBtpEngine, HandleCharacteristicReceivedTwoPacket)
     EXPECT_EQ(mBtpEngine.RxState(), BtpEngine::kState_InProgress);
 
     auto packet1 = System::PacketBufferHandle::NewWithData(packetData1, sizeof(packetData1));
-    EXPECT_EQ(packet1->DataLength(), 3);
+    EXPECT_EQ(packet1->DataLength(), static_cast<size_t>(3));
 
     EXPECT_EQ(mBtpEngine.HandleCharacteristicReceived(std::move(packet1), receivedAck, didReceiveAck), CHIP_NO_ERROR);
     EXPECT_EQ(mBtpEngine.RxState(), BtpEngine::kState_Complete);
@@ -102,7 +103,7 @@ TEST_F(TestBtpEngine, HandleCharacteristicReceivedThreePacket)
     constexpr uint8_t packetData2[] = { to_underlying(BtpEngine::HeaderFlags::kEndMessage), 0x03, 0xff };
 
     auto packet0 = System::PacketBufferHandle::NewWithData(packetData0, sizeof(packetData0));
-    EXPECT_EQ(packet0->DataLength(), 5);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(5));
 
     SequenceNumber_t receivedAck;
     bool didReceiveAck;
@@ -110,13 +111,13 @@ TEST_F(TestBtpEngine, HandleCharacteristicReceivedThreePacket)
     EXPECT_EQ(mBtpEngine.RxState(), BtpEngine::kState_InProgress);
 
     auto packet1 = System::PacketBufferHandle::NewWithData(packetData1, sizeof(packetData1));
-    EXPECT_EQ(packet1->DataLength(), 3);
+    EXPECT_EQ(packet1->DataLength(), static_cast<size_t>(3));
 
     EXPECT_EQ(mBtpEngine.HandleCharacteristicReceived(std::move(packet1), receivedAck, didReceiveAck), CHIP_NO_ERROR);
     EXPECT_EQ(mBtpEngine.RxState(), BtpEngine::kState_InProgress);
 
     auto packet2 = System::PacketBufferHandle::NewWithData(packetData2, sizeof(packetData2));
-    EXPECT_EQ(packet2->DataLength(), 3);
+    EXPECT_EQ(packet2->DataLength(), static_cast<size_t>(3));
 
     EXPECT_EQ(mBtpEngine.HandleCharacteristicReceived(std::move(packet2), receivedAck, didReceiveAck), CHIP_NO_ERROR);
     EXPECT_EQ(mBtpEngine.RxState(), BtpEngine::kState_Complete);
@@ -133,7 +134,7 @@ TEST_F(TestBtpEngine, HandleCharacteristicSendOnePacket)
 
     EXPECT_TRUE(mBtpEngine.HandleCharacteristicSend(packet0.Retain(), false));
     EXPECT_EQ(mBtpEngine.TxState(), BtpEngine::kState_Complete);
-    EXPECT_EQ(packet0->DataLength(), 5);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(5));
 }
 
 TEST_F(TestBtpEngine, HandleCharacteristicSendTwoPacket)
@@ -147,11 +148,11 @@ TEST_F(TestBtpEngine, HandleCharacteristicSendTwoPacket)
 
     EXPECT_TRUE(mBtpEngine.HandleCharacteristicSend(packet0.Retain(), false));
     EXPECT_EQ(mBtpEngine.TxState(), BtpEngine::kState_InProgress);
-    EXPECT_EQ(packet0->DataLength(), 20);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(20));
 
     EXPECT_TRUE(mBtpEngine.HandleCharacteristicSend(nullptr, false));
     EXPECT_EQ(mBtpEngine.TxState(), BtpEngine::kState_Complete);
-    EXPECT_EQ(packet0->DataLength(), 16);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(16));
 }
 
 // Send 40-byte payload.
@@ -169,15 +170,15 @@ TEST_F(TestBtpEngine, HandleCharacteristicSendThreePacket)
 
     EXPECT_TRUE(mBtpEngine.HandleCharacteristicSend(packet0.Retain(), false));
     EXPECT_EQ(mBtpEngine.TxState(), BtpEngine::kState_InProgress);
-    EXPECT_EQ(packet0->DataLength(), 20);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(20));
 
     EXPECT_TRUE(mBtpEngine.HandleCharacteristicSend(nullptr, false));
     EXPECT_EQ(mBtpEngine.TxState(), BtpEngine::kState_InProgress);
-    EXPECT_EQ(packet0->DataLength(), 20);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(20));
 
     EXPECT_TRUE(mBtpEngine.HandleCharacteristicSend(nullptr, false));
     EXPECT_EQ(mBtpEngine.TxState(), BtpEngine::kState_Complete);
-    EXPECT_EQ(packet0->DataLength(), 8);
+    EXPECT_EQ(packet0->DataLength(), static_cast<size_t>(8));
 }
 
 } // namespace
